Cleaned up the code, made it neater and shorter.

+++ trunk/AdhanAlarm/src/islam/adhanalarm/WakeUpAndDoSomething.java
@@ -8,7 +8,7 @@ public class WakeUpAndDoSomething extends BroadcastReceiver {
 	@Override
 	public void onReceive(Context context, Intent intent) {
         Intent i = new Intent(context, AdhanAlarm.class);
-        i.putExtra("nextNotificationTime", intent.getIntExtra("nextNotificationTime", -1));
+        i.putExtra("nextNotificationTime", intent.getShortExtra("nextNotificationTime", (short)-1));
         i.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK);
         context.startActivity(i); // Simply calls the AdhanAlarm activity which may play an alarm in onResume
     }
